+++ components/index.js
@@ -11,9 +11,7 @@ export { default as BrokerPortalPage } from './Pages/BrokerPortal';
 export { default as ServiceUserDetails } from './Pages/BrokerPortal/ServiceUserDetails';
 export { default as CareDetails } from './Pages/ServiceUser/Packages/CareDetails';
 export { default as PackageRequest } from './Pages/ServiceUser/Packages/PackageRequest';
-export { default as BrokerageBorderCost } from './Pages/CarePackages/BrokerageBorderCost';
 export { default as ReviewPackageDetails } from './Pages/CarePackages/ReviewPackageDetails';
-export { default as SearchServiceUser } from './Pages/BrokerReferral/SearchForServiceUser';
 export { default as NewCareChargePackages } from './Pages/CareCharges/NewCareChargePackages';
 export { default as PackageApprovals } from './Pages/PackageApprovals';
 export { default as DynamicBreadcrumbs } from './Pages/DynamicBreadcrumbs';
